+++ src/main/java/com/gruppe4b/edivator/backend/service/ImageEditService.java
@@ -15,7 +15,7 @@ public interface ImageEditService {
     public String flip(String imageId,boolean horizontal);
     public String turnLeft(String imageId);
     public String turnRight(String imageId);
-    public String crop(String imageId, int leftBorder, int rightBorder, int topBorder, int bottomBorder);
+    public String crop(String imageId, float leftBorder, float rightBorder, float topBorder, float bottomBorder);
     public String applyLuckyFilter(String imageId);
 
 }
