+++ src/main/java/com/gruppe4b/edivator/backend/service/ImageEditServiceImpl.java
@@ -6,6 +6,7 @@ import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.nio.channels.FileChannel;
 
+import com.google.appengine.api.datastore.TransactionOptions;
 import com.google.appengine.api.images.Image;
 import com.google.appengine.api.images.ImagesService;
 import com.google.appengine.api.images.ImagesServiceFactory;
@@ -153,7 +154,7 @@ public class ImageEditServiceImpl implements ImageEditService {
         return gson.toJson(createResponse(url,id));
     }
 
-    public String crop(String imageId, int leftBorder, int rightBorder, int topBorder, int bottomBorder) {
+    public String crop(String imageId, float leftBorder, float rightBorder, float topBorder, float bottomBorder) {
         Image croppingImage = imageStoreService.getImageFromCloudStorage(imageId);
 
         Transform transform = ImagesServiceFactory.makeCrop(leftBorder,topBorder,rightBorder,bottomBorder);
